Inject services in page

+++ lib/pages/accounts.dart
@@ -1,24 +1,21 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:reef_mobile_app/model/ReefState.dart';
 import 'package:reef_mobile_app/model/StorageKey.dart';
 import 'package:reef_mobile_app/model/account/stored_account.dart';
 import 'package:reef_mobile_app/service/JsApiService.dart';
 import 'package:reef_mobile_app/service/StorageService.dart';
 
 class AccountPage extends StatefulWidget {
-  const AccountPage({super.key});
+  AccountPage(this.jsApiService, this.storageService);
+  final JsApiService jsApiService;
+  final StorageService storageService;
 
   @override
   State<AccountPage> createState() => _AccountPageState();
 }
 
 class _AccountPageState extends State<AccountPage> {
-  final storage = StorageService();
-  final jsApiService = JsApiService();
-  late ReefState reefState;
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -35,11 +32,11 @@ class _AccountPageState extends State<AccountPage> {
                 onPressed: generateAccount,
                 child: const Text('Generate Account')),
             TextButton(
-                onPressed: _getAccounts, child: const Text('Get Accounts')),
+                onPressed: _callSaveAccount, child: const Text('Add Account')),
             TextButton(
-                onPressed: _addAccount, child: const Text('Add Account')),
+                onPressed: getAccount, child: const Text('Get Accounts')),
             TextButton(
-                onPressed: _deleteAccount, child: const Text('Delete Account')),
+                onPressed: deleteAccount, child: const Text('Delete Account')),
           ],
         ),
       ),
@@ -47,77 +44,71 @@ class _AccountPageState extends State<AccountPage> {
   }
 
   void initWasm() async {
-    var res = await jsApiService.jsPromise('keyring.callCryptoWaitReady()');
+    var res = await widget.jsApiService.jsPromise('keyring.callCryptoWaitReady()');
     print("res: ${res}");
   }
 
   void test() async {
     const mnemonic =
-        'abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about';
+        'debris pink stairs furnace rescue toddler face finger vast trash repair bone';
     checkMnemonicValid(mnemonic);
-    addressFromMnemonic(mnemonic);
+    // addressFromMnemonic(mnemonic);
   }
 
   void generateAccount() async {
-    var res = await jsApiService.jsPromise('keyring.gen()');
+    var res = await widget.jsApiService.jsPromise('keyring.gen()');
     print("res: ${res}");
   }
 
   void checkMnemonicValid(String mnemonic) async {
-    var res =
-        await jsApiService.jsCall('keyring.checkMnemonicValid(${mnemonic})');
-    print("res: ${res}");
+    var valid = await widget.jsApiService
+        .jsCall('keyring.checkMnemonicValid("${mnemonic}")');
+    print("valid: ${valid}");
   }
 
+  /// Returns the address of the account with the given mnemonic 
   void addressFromMnemonic(String mnemonic) async {
-    var res =
-        await jsApiService.jsCall('keyring.addressFromMnemonic(${mnemonic})');
-    print("res: ${res}");
+    var address = await widget.jsApiService
+        .jsCall('keyring.addressFromMnemonic("${mnemonic}")');
+    print("address: ${address}");
   }
 
-  void _getAccounts() async {
-    final accounts = await getAccounts();
-    print("accounts number: ${accounts.length}");
-    if (accounts.isNotEmpty) {
-      print("account 1: ${accounts[0].address}");
-    }
-  }
 
-  void _addAccount() {
-    addAccount(
+  void _callSaveAccount() {
+    saveAccount(
         "debris pink stairs furnace rescue toddler face finger vast trash repair bone",
         "5HgkDz5N1L1PvwSZVDkdZ2fgBAWu6kAHYruH1QvCL3DvwRQj",
         "<svg></svg>");
   }
 
-  void _deleteAccount() async {
-    final accounts = await getAccounts();
-    deleteAccount(accounts[0]);
-  }
-
-  Future addAccount(String mnemonic, String address, String svg) async {
+  /// Save account to storage
+  Future saveAccount(String mnemonic, String address, String svg) async {
     final account = StoredAccount()
       ..mnemonic = mnemonic
       ..address = address
       ..svg = svg;
 
-    print(account);
-    var res = await storage.setValue(StorageKey.stored_accounts.name, account);
-    print("res: ${res}");
+    await widget.storageService.setValue(StorageKey.account.name, account);
+    print("Saved account ${account.address}.");
   }
 
-  void deleteAccount(StoredAccount account) {
-    account.delete();
+  /// Get all account from storage
+  Future<StoredAccount?> getAccount() async {
+    var account = await widget.storageService.getValue(StorageKey.account.name);
+    if (account == null) {
+      print("No account found.");
+      return null;
+    }
+    print("Fetched account ${account.address}.");
+    return account;
   }
 
-  Future<List<StoredAccount>> getAccounts() async {
-    // final box = Hive.box<StoredAccount>('accounts');
-    // return box.values.toList().cast<StoredAccount>();
-    print("getAccounts");
-    var accounts = await storage.getValue(StorageKey.stored_accounts.name);
-    print(accounts);
-    var acc = accounts.toList().cast<StoredAccount>();
-    print(acc);
-    return acc;
+  /// Delete account from storage
+  void deleteAccount() async {
+    var account = await getAccount();
+    if (account != null) {
+      account.delete();
+      print("Deleted accont ${account.address}.");
+    }
   }
 }
